First run at a professional statement
As per the module, I downloaded the outline and did my best to follow it. The statement is something that will evolve over time and I'll be adding each of the updates as separate files and showcasing the evolution of what I believe me professional statement is going to be!